* Port [3877] to trunk from 1.2.16  * Cleanup all time_t warnings
svn: 3878

+++ src/auth.c
@@ -257,7 +257,7 @@ static int auth_tell_walk(Auth *auth, void *data)
 
 //  dprintf(idx, "%s(%s!%s) [%s] authtime: %li, atime: %li, Status: %d\n", auth->bd ? "x " : "", auth->nick, 
   dprintf(idx, "(%s!%s) [%s] authtime: %li, atime: %li, Status: %d\n", auth->nick, 
-        auth->host, auth->handle, auth->authtime, auth->atime, auth->Status());
+        auth->host, auth->handle, (long)auth->authtime, (long)auth->atime, auth->Status());
   
   return 0;
 }
